Initial steps for reviving of the F# support
