+++ src/Stryker.Core/Stryker.Core/Mutants/BaseMutantOrchestrator.cs
@@ -28,6 +28,10 @@ namespace Stryker.Core.Mutants
         }
     }
 
+    /// <summary>
+    /// Mutates abstract syntax trees using mutators and places all mutations inside the abstract syntax tree.
+    /// Orchestrator: to arrange or manipulate, especially by means of clever or thorough planning or maneuvering.
+    /// </summary>
     public abstract class BaseMutantOrchestrator<T> : BaseMutantOrchestrator
     {
         protected BaseMutantOrchestrator(StrykerOptions input) : base(input)
