+++ src/Stryker.Core/Stryker.Core/Mutants/FsharpOrchestrators/DefaultOrchestrator.cs
@@ -1,0 +1,10 @@
+﻿namespace Stryker.Core.Mutants.FsharpOrchestrators
+{
+    public class DefaultOrchestrator<T> : IFsharpTypeHandler<T>
+    {
+        public virtual T Mutate(T input, FsharpMutantOrchestrator iterator)
+        {
+            return input;
+        }
+    }
+}
