+++ src/Stryker.Core/Stryker.Core/Mutants/FsharpOrchestrators/IFsharpTypehandle.cs
@@ -1,0 +1,7 @@
+namespace Stryker.Core.Mutants.FsharpOrchestrators
+{
+    public interface IFsharpTypeHandler<T>
+    {
+        public T Mutate(T input, FsharpMutantOrchestrator iterator);
+    }
+}
