+++ src/Stryker.Core/Stryker.Core/Mutants/FsharpOrchestrators/NestedModuleOrchestrator.cs
@@ -1,0 +1,16 @@
+﻿using static FSharp.Compiler.SyntaxTree;
+
+namespace Stryker.Core.Mutants.FsharpOrchestrators
+{
+    public class NestedModuleOrchestrator : IFsharpTypeHandler<SynModuleDecl>
+    {
+        public SynModuleDecl Mutate(SynModuleDecl input, FsharpMutantOrchestrator iterator)
+        {
+            var castinput = input as SynModuleDecl.NestedModule;
+
+            var visitedDeclarations = iterator.Mutate(castinput.decls);
+            return SynModuleDecl.NewNestedModule(castinput.moduleInfo, castinput.isRecursive, visitedDeclarations, castinput.isContinuing, castinput.range);
+        }
+    }
+
+}
