+++ src/Stryker.Core/Stryker.Core/ProjectComponents/Fsharp/FsharpFolderComposite.cs
@@ -2,7 +2,7 @@ using static FSharp.Compiler.SyntaxTree;
 
 namespace Stryker.Core.ProjectComponents
 {
-    public class FsharpFolderComposite : FolderComposite<ParsedInput>
+    public class FsharpFolderComposite : FolderComposite<ParsedInput>, IReadOnlyFolderComposite
     {
 
     }
